feat: add component ListEvents
