+++ src/pages/index.css
@@ -63,6 +63,10 @@
   width: calc(100% - 256px);
 }
 
+.ui.container .events__all-events-title {
+  margin-top: 60px;
+}
+
 @media only screen and (max-width: 768px) {
   .ui.container .Row .left.Column.sidebar {
     display: none;
